requirements.txt
phidata
python-dotenv
groq
fastapi
yfinance
packaging
uvicorn
duckduckgo-search
openai